chore: made config easier to navigate
