--- src/config/index.ts
@@ -1,5 +1,0 @@
-import { socials } from './socials';
-
-export {
-  socials
-};
